+++ db/HeartbeatData.js
@@ -1,9 +1,11 @@
 class HeartbeatData {
   constructor(streamEntry) {
     this.timestamp = streamEntry[0].split('-')[0] // get timestamp in millis from redis ID
-    this.doorStatus = streamEntry[1][1]
-    this.doorTime = streamEntry[1][3]
-    this.insTime = streamEntry[1][5]
+    this.doorMissedMessagesCount = streamEntry[1][1]
+    this.doorLowBatteryFlag = streamEntry[1][3]
+    this.doorTimeSinceLastHeartbeat = streamEntry[1][5]
+    this.resetReason = streamEntry[1][7]
+    this.stateTransitionsArray = streamEntry[1][9]
   }
 }