+++ db/db.js
@@ -105,7 +105,7 @@ async function createLocationFromRow(r, clientParam) {
     const client = createClientFromRow(results.rows[0])
 
     // prettier-ignore
-    return new Location(r.locationid, r.display_name, r.movement_threshold, r.duration_timer, r.stillness_timer, r.heartbeat_sent_alerts, r.heartbeat_alert_recipients, r.door_particlecoreid, r.radar_particlecoreid, r.radar_type, r.reminder_timer, r.fallback_timer, r.twilio_number, r.fallback_phonenumbers, r.initial_timer, r.is_active, r.firmware_state_machine, r.siren_particle_id, client)
+    return new Location(r.locationid, r.display_name, r.movement_threshold, r.duration_timer, r.stillness_timer, r.heartbeat_sent_alerts, r.heartbeat_alert_recipients, r.door_particlecoreid, r.radar_particlecoreid, r.radar_type, r.reminder_timer, r.fallback_timer, r.twilio_number, r.fallback_phonenumbers, r.initial_timer, r.is_active, r.firmware_state_machine, r.siren_particle_id, r.sent_low_battery_alert_at, client)
   } catch (err) {
     helpers.log(err.toString())
   }
@@ -312,6 +312,19 @@ async function updateSentAlerts(locationid, sentalerts, clientParam) {
     }
 
     return await createLocationFromRow(results.rows[0], clientParam)
+  } catch (err) {
+    helpers.log(err.toString())
+  }
+}
+
+async function updateLowBatteryAlertTime(locationid, clientParam) {
+  try {
+    await runQuery(
+      'updateLowBatteryAlertTime',
+      'UPDATE locations SET sent_low_battery_alert_at = NOW() WHERE locationid = $1',
+      [locationid],
+      clientParam,
+    )
   } catch (err) {
     helpers.log(err.toString())
   }
@@ -699,11 +712,11 @@ async function createLocationFromBrowserForm(locationid, displayName, doorCoreId
 
 // Adds a location table entry
 // eslint-disable-next-line prettier/prettier
-async function createLocation(locationid, movementThreshold, stillnessTimer, durationTimer, reminderTimer, initialTimer, heartbeatAlertRecipients, twilioNumber, fallbackNumbers, fallbackTimer, displayName, doorCoreId, radarCoreId, radarType, isActive, firmwareStateMachine, sirenParticleId, clientId, clientParam) {
+async function createLocation(locationid, movementThreshold, stillnessTimer, durationTimer, reminderTimer, initialTimer, heartbeatAlertRecipients, twilioNumber, fallbackNumbers, fallbackTimer, displayName, doorCoreId, radarCoreId, radarType, isActive, firmwareStateMachine, sirenParticleId, sentLowBatteryAlertAt, clientId, clientParam) {
   try {
     await runQuery(
       'createLocation',
-      'INSERT INTO locations(locationid, movement_threshold, stillness_timer, duration_timer, reminder_timer, initial_timer, heartbeat_alert_recipients, twilio_number, fallback_phonenumbers, fallback_timer, display_name, door_particlecoreid, radar_particlecoreid, radar_type, is_active, firmware_state_machine, siren_particle_id, client_id) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18)',
+      'INSERT INTO locations(locationid, movement_threshold, stillness_timer, duration_timer, reminder_timer, initial_timer, heartbeat_alert_recipients, twilio_number, fallback_phonenumbers, fallback_timer, display_name, door_particlecoreid, radar_particlecoreid, radar_type, is_active, firmware_state_machine, siren_particle_id, sent_low_battery_alert_at, client_id) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19)',
       [
         locationid,
         movementThreshold,
@@ -722,6 +735,7 @@ async function createLocation(locationid, movementThreshold, stillnessTimer, dur
         isActive,
         firmwareStateMachine,
         sirenParticleId,
+        sentLowBatteryAlertAt,
         clientId,
       ],
       clientParam,
@@ -880,4 +894,5 @@ module.exports = {
   rollbackTransaction,
   getCurrentTime,
   createLocationFromBrowserForm,
+  updateLowBatteryAlertTime,
 }