+++ test/integration/BraveAlerterConfiguratorTest/getAlertSessionByPhoneNumberTest.js
@@ -38,6 +38,7 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSessionByPhoneN
       true,
       false,
       null,
+      '2021-03-09T19:37:28.176Z',
       client.id,
     )
     const locationId = (await db.getLocations())[0].locationid