+++ test/integration/dashboardTest/submitEditLocationTest.js
@@ -48,6 +48,7 @@ describe('dashboard.js integration tests: submitEditLocation', () => {
       true,
       false,
       null,
+      '2021-03-09T19:37:28.176Z',
       this.client.id,
     )