+++ test/integration/sirenTest/handleSirenEscalatedTest.js
@@ -27,6 +27,7 @@ const MOVEMENT_THRESHOLD = 40
 const INITIAL_TIMER = 1
 const STILLNESS_TIMER = 1.5
 const DURATION_TIMER = 3
+const firstLowBatteryAlert = '2021-03-09T19:37:28.176Z'
 
 async function sirenEscalatedAlert(coreID) {
   let response
@@ -72,6 +73,7 @@ describe('siren.js integration tests: handleSirenEscalated', () => {
       true,
       true,
       testSirenId,
+      firstLowBatteryAlert,
       client.id,
     )
     sandbox.stub(braveAlerter, 'startAlertSession')